requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
diskcache==5.6.3
einops==0.7.0
filelock==3.13.1
fsspec==2024.3.0
gguf==0.6.0
huggingface-hub==0.21.4
idna==3.6
Jinja2==3.1.3
llama_cpp_python==0.2.56
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
packaging==24.0
protobuf==4.25.3
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sentencepiece==0.1.99
sympy==1.12
tokenizers==0.15.2
torch==2.1.2
tqdm==4.66.2
transformers==4.38.2
triton==2.1.0
typing_extensions==4.10.0
urllib3==2.2.1
